+++ wagtail/tests/utils/__init__.py
@@ -1,0 +1,2 @@
+from .page_tests import *  # NOQA
+from .wagtail_tests import *  # NOQA
